requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Babel==2.15.0
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-admin-autocomplete-filter==0.7.1
django-admin-lightweight-date-hierarchy==1.3.0
django-admin-list-filter-dropdown==1.0.3
django-admin-rangefilter==0.12.4
django-admin-sortable2==2.1.10
django-autocomplete-light==3.11.0
django-baton==3.1.0
django-cors-headers==4.3.1
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.2
django-flatpickr==2.0.3
django-import-export==3.3.8
django-js-asset==2.2.0
django-mptt==0.16.0
django-mptt-admin==2.6.2
django-nested-admin==4.0.2
django-reverse-admin==2.9.6
django-tinymce==4.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
idna==3.7
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
pandas==2.2.2
pillow==10.3.0
prettytable==3.10.0
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
pydot==2.0.0
PyJWT==2.8.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-monkey-business==1.0.0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
wcwidth==0.2.13
xlrd==2.0.1
xlwt==1.3.0
